Move actions stores to /action module
