+++ apps/web/modules/action/index.ts
@@ -1,1 +1,3 @@
+export { ActionsStore } from './actions-store';
+export { ActionsStoreProvider, useActionsStoreContext } from './actions-store-provider';
 export * as Action from './action';
