+++ apps/web/modules/stores/triple-store.test.tsx
@@ -1,8 +1,8 @@
 import { when } from '@legendapp/state';
 import { describe, expect, it } from 'vitest';
+import { ActionsStore } from '../action';
 import { makeStubTriple, MockNetwork } from '../services/mock-network';
 import { TripleStore } from './triple-store';
-import { ActionsStore } from './actions-store';
 
 describe('TripleStore', () => {
   it('Initializes to empty', async () => {
